LayoutTests/imported/w3c/web-platform-tests/css/cssom-view/Element-currentCSSZoom-expected.txt


FAIL Element.currentCSSZoom basic test assert_equals: Unzoomed content expected (number) 1 but got (undefined) undefined
FAIL Element.currentCSSZoom reacts to style changes assert_equals: currentCSSZoom reacts to style changes expected (number) 2 but got (undefined) undefined
